backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
drf-yasg==1.21.7
google-generativeai==0.3.1
Pillow==10.1.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.26
python-jose==3.3.0
